api/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-multipart
logging